+++ CRUD_demo/services/AdminService.php
@@ -13,16 +13,15 @@ class AdminService
     }
     public function login(LoginRequest $request)
     {
-        $user = $this->adminRepo->findByEmail($request->getEmail());
-        if (empty($user)) {
-            throw new Exception("Sai email hoặc mật khẩu");
+        $admin = $this->adminRepo->findByEmail($request->getEmail());
+        if (empty($admin)) {
+            throw new Exception("Email ko tồn tại");
         }
-        if (password_verify($request->getPassword(), $user->getPassword())) {
-            return $user;
+        if (password_verify($request->getPassword(), $admin->getPassword())) {
+            return $admin;
         } else {
             throw new Exception("Sai email hoặc mật khẩu");
         }
-
     }
     public function getAdminById($id)
     {
@@ -33,20 +32,30 @@ class AdminService
         return $this->adminRepo->getAll();
     }
 
+    public function search(SearchRequest $request)
+    {
+        $data = $request->toArray();
+        return $this->adminRepo->search($data);
+    }
+
     public function createAdmin(AdminCreateRequest $admin)
     {
         $data = $admin->toArray();
         $errors = [];
-        if (strlen(empty($data['password'])) || !Validation::validatePassword($data['password'])) {
+        if (strlen(empty($data['password'])) || !Validation::validatePassword(trim($data['password']))) {
             $errors['passwordError'] = "Mật khẩu ko hợp lệ";
+        } else {
+            $data["password"] = password_hash(trim($data['password']), PASSWORD_DEFAULT);
         }
-        if (strlen(empty($data['name'])) || strlen($data['name']) > 128) {
+        if (strlen(empty($data['name'])) || strlen(trim($data['name'])) > 128) {
             $errors['nameError'] = "Tên phải từ 1 đến 128 ký tự";
         }
-        if (strlen(empty($data['email'])) || strlen($data['email']) > 128 || !Validation::validateEmail($data['email'])) {
-            $errors['emailError'] = "Email phải từ 1 đến 128 ký tự";
+        if (strlen(empty($data['email'])) || strlen(trim($data['email'])) > 128 || !Validation::validateEmail(trim($data['email']))) {
+            $errors['emailError'] = "Email ko hợp lệ";
+        } else if ($this->adminRepo->existedEmail(trim($data['email']))) {
+            $errors['emailError'] = "Email đã tồn tại";
         }
-        if (strlen(empty($data['avatar'])) || strlen($data['avatar']) > 128) {
+        if (strlen(empty($data['avatar'])) || strlen(trim($data['avatar'])) > 128) {
             $errors['avatarError'] = "Avatar phải từ 1 đến 128 ký tự";
         }
         if (!empty($errors)) {
@@ -58,13 +67,18 @@ class AdminService
     {
         $data = $admin->toArray();
         $errors = [];
-        if (strlen(empty($data['name'])) || strlen($data['name']) > 128) {
+        if (strlen(empty($data['name'])) || strlen(trim($data['name'])) > 128) {
             $errors['nameError'] = "Tên phải từ 1 đến 128 ký tự";
         }
-        if (strlen(empty($data['email'])) || strlen($data['email'] || Validation::validateEmail($data['email'])) > 128) {
-            $errors['emailError'] = "Email phải từ 1 đến 128 ký tự";
+        if (strlen(empty($data['email'])) || strlen(trim($data['email'])) > 128 || !Validation::validateEmail(trim($data['email']))) {
+            $errors['emailError'] = "Email ko hợp lệ";
+        } else {
+            $checkedId = $this->adminRepo->existedEmail(trim($data['email']));
+            if ($checkedId != $id) {
+                $errors['emailError'] = "Email đã tồn tại";
+            }
         }
-        if (strlen(empty($data['avatar'])) || strlen($data['avatar']) > 128) {
+        if (strlen(empty($data['avatar'])) || strlen(trim($data['avatar'])) > 128) {
             $errors['avatarError'] = "Avatar phải từ 1 đến 128 ký tự";
         }
         if (!empty($errors)) {
@@ -74,6 +88,10 @@ class AdminService
     }
     public function deleteAdmin($id)
     {
+        $user = $this->adminRepo->findById($id);
+        if (empty($user)) {
+            throw new Exception("not found");
+        }
         return $this->adminRepo->delete($id);
     }
 }
